Removed a boatload of old movement code. Fixed a few code warnings. Added more robust movement checking with our Map GridCube checking function.

+++ trunk/SpaceHaste/SpaceHaste/SpaceHaste/GameObjects/GameObject.cs
@@ -26,7 +26,7 @@ namespace SpaceHaste.GameObjects
         public double NeededEnergy;
         public GridCube location;
 
-        public float MovementRange;
+        public int MovementRange;
 
         //Model/Render Info
         public float Scale;
